chore: Update some dependencies
Cannot update to Svelte 5 right now, because it requires major refactoring due to changed handling of state.
Cannot upgrade to Vite 6 right now, because it requires upgrade of @sveltejs/vite-plugin-svelte for compatibility.
Cannot upgrade @sveltejs/vite-plugin-svelte to 4.x.x, because it likely requires a newer version of Svelte.

What a nice loop?!